Refoctor, uniform MySQL/POSTGRESQL/MARIADB plugin config to a single one
