+++ apm-sniffer/apm-sdk-plugin/jdbc-commons/src/main/java/org/apache/skywalking/apm/plugin/jdbc/JDBCPluginConfig.java
@@ -23,59 +23,12 @@ import org.apache.skywalking.apm.agent.core.boot.PluginConfig;
 public class JDBCPluginConfig {
     public static class Plugin {
         @PluginConfig(root = JDBCPluginConfig.class)
-        public static class MySQL {
+        public static class JDBC {
             /**
              * If set to true, the parameters of the sql (typically {@link java.sql.PreparedStatement}) would be
              * collected.
              */
             public static boolean TRACE_SQL_PARAMETERS = false;
-            /**
-             * For the sake of performance, SkyWalking won't save the entire parameters string into the tag, but only
-             * the first {@code SQL_PARAMETERS_MAX_LENGTH} characters.
-             * <p>
-             * Set a negative number to save the complete parameter string to the tag.
-             */
-            public static int SQL_PARAMETERS_MAX_LENGTH = 512;
-            /**
-             * For the sake of performance, SkyWalking won't save the entire sql body into the tag, but only the first
-             * {@code SQL_BODY_MAX_LENGTH} characters.
-             * <p>
-             * Set a negative number to save the complete sql body to the tag.
-             */
-            public static int SQL_BODY_MAX_LENGTH = 2048;
-        }
-
-        @PluginConfig(root = JDBCPluginConfig.class)
-        public static class POSTGRESQL {
-            /**
-             * If set to true, the parameters of the sql (typically {@link java.sql.PreparedStatement}) would be
-             * collected.
-             */
-            public static boolean TRACE_SQL_PARAMETERS = false;
-
-            /**
-             * For the sake of performance, SkyWalking won't save the entire parameters string into the tag, but only
-             * the first {@code SQL_PARAMETERS_MAX_LENGTH} characters.
-             * <p>
-             * Set a negative number to save the complete parameter string to the tag.
-             */
-            public static int SQL_PARAMETERS_MAX_LENGTH = 512;
-            /**
-             * For the sake of performance, SkyWalking won't save the entire sql body into the tag, but only the first
-             * {@code SQL_BODY_MAX_LENGTH} characters.
-             * <p>
-             * Set a negative number to save the complete sql body to the tag.
-             */
-            public static int SQL_BODY_MAX_LENGTH = 2048;
-        }
-
-        public static class MARIADB {
-            /**
-             * If set to true, the parameters of the sql (typically {@link java.sql.PreparedStatement}) would be
-             * collected.
-             */
-            public static boolean TRACE_SQL_PARAMETERS = false;
-
             /**
              * For the sake of performance, SkyWalking won't save the entire parameters string into the tag, but only
              * the first {@code SQL_PARAMETERS_MAX_LENGTH} characters.
