+++ apm-sniffer/apm-sdk-plugin/mariadb-2.x-plugin/src/test/java/org/apache/skywalking/apm/plugin/jdbc/mariadb/v2/StatementExecuteMethodsInterceptorTest.java
@@ -99,7 +99,7 @@ public class StatementExecuteMethodsInterceptorTest {
 
     @Test
     public void testExecuteStatementWithLimitSqlBody() {
-        JDBCPluginConfig.Plugin.MARIADB.SQL_BODY_MAX_LENGTH = 10;
+        JDBCPluginConfig.Plugin.JDBC.SQL_BODY_MAX_LENGTH = 10;
         serviceMethodInterceptor.beforeMethod(objectInstance, method, new Object[]{SQL}, null, null);
         serviceMethodInterceptor.afterMethod(objectInstance, method, new Object[]{SQL}, null, null);
