+++ apm-sniffer/apm-sdk-plugin/mysql-common/src/main/java/org/apache/skywalking/apm/plugin/jdbc/mysql/PreparedStatementExecuteMethodsInterceptor.java
@@ -28,7 +28,7 @@ import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceM
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
 import org.apache.skywalking.apm.plugin.jdbc.JDBCPluginConfig;
 import org.apache.skywalking.apm.plugin.jdbc.PreparedStatementParameterBuilder;
-import org.apache.skywalking.apm.plugin.jdbc.SqlBodyBuilder;
+import org.apache.skywalking.apm.plugin.jdbc.SqlBodyUtil;
 import org.apache.skywalking.apm.plugin.jdbc.define.StatementEnhanceInfos;
 import org.apache.skywalking.apm.plugin.jdbc.trace.ConnectionInfo;
 
@@ -54,11 +54,10 @@ public class PreparedStatementExecuteMethodsInterceptor implements InstanceMetho
                     .getStatementName()), connectInfo.getDatabasePeer());
             Tags.DB_TYPE.set(span, "sql");
             Tags.DB_INSTANCE.set(span, connectInfo.getDatabaseName());
-            Tags.DB_STATEMENT.set(span, new SqlBodyBuilder().setMaxLength(JDBCPluginConfig.Plugin.MySQL.SQL_BODY_MAX_LENGTH)
-                    .setSqlBody(cacheObject.getSql()).build());
+            Tags.DB_STATEMENT.set(span, SqlBodyUtil.limitSqlBodySize(cacheObject.getSql()));
             span.setComponent(connectInfo.getComponent());
 
-            if (JDBCPluginConfig.Plugin.MySQL.TRACE_SQL_PARAMETERS) {
+            if (JDBCPluginConfig.Plugin.JDBC.TRACE_SQL_PARAMETERS) {
                 final Object[] parameters = cacheObject.getParameters();
                 if (parameters != null && parameters.length > 0) {
                     int maxIndex = cacheObject.getMaxIndex();
@@ -98,7 +97,6 @@ public class PreparedStatementExecuteMethodsInterceptor implements InstanceMetho
         return new PreparedStatementParameterBuilder()
             .setParameters(parameters)
             .setMaxIndex(maxIndex)
-            .setMaxLength(JDBCPluginConfig.Plugin.MySQL.SQL_PARAMETERS_MAX_LENGTH)
             .build();
     }
 }
