+++ apm-sniffer/apm-sdk-plugin/mysql-common/src/main/java/org/apache/skywalking/apm/plugin/jdbc/mysql/StatementExecuteMethodsInterceptor.java
@@ -25,8 +25,7 @@ import org.apache.skywalking.apm.agent.core.context.trace.SpanLayer;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceMethodsAroundInterceptor;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
-import org.apache.skywalking.apm.plugin.jdbc.JDBCPluginConfig;
-import org.apache.skywalking.apm.plugin.jdbc.SqlBodyBuilder;
+import org.apache.skywalking.apm.plugin.jdbc.SqlBodyUtil;
 import org.apache.skywalking.apm.plugin.jdbc.define.StatementEnhanceInfos;
 import org.apache.skywalking.apm.plugin.jdbc.trace.ConnectionInfo;
 
@@ -59,8 +58,7 @@ public class StatementExecuteMethodsInterceptor implements InstanceMethodsAround
             String sql = "";
             if (allArguments.length > 0) {
                 sql = (String) allArguments[0];
-                sql = new SqlBodyBuilder().setMaxLength(JDBCPluginConfig.Plugin.MySQL.SQL_BODY_MAX_LENGTH)
-                        .setSqlBody(sql).build();
+                sql = SqlBodyUtil.limitSqlBodySize(sql);
             }
             Tags.DB_STATEMENT.set(span, sql);
             span.setComponent(connectInfo.getComponent());
